// backend/src/entities/enums/user.document.type.ts
export enum UserDocumentTypes {
  CITIES_ID = 'cities_id',
  DRVING_LICENSE = 'driving_license',
  PASSPORT = 'passport',

  ENTRY_STAMP = 'entery_stamp',
  GCC_ID = 'gcc_id',
  DRVING_LICENSE_HOME_COUNTRY = 'driving_license_home_country',
  DRVING_LICENSE_GCC = 'driving_license_gcc',
  INTERNATIONAL_DRIVING_LICENSE = 'international_driving_license',
  TRANSLATION_OF_DRIVING_LICENSE = 'translation_of_driving_license',
  TOURIST_VISA = 'tourist_visa',
  OTHER = 'other'
}